core: clear the -Wall -Wextra warnings in cpu, memory, debug, scheduler
Behaviour-neutral:
- sign-compare: (int)page < g_page_count, where both sides are uint32_t
  (the int was converted straight back); the casts go.
- dead code deleted: disasm_cache_scope, uint128_shr/uint128_shl,
  load_png_to_framebuffer (165 lines, no callers), unused locals and a
  never-targeted done: label in the disassembler epilogue, a
  saved-and-never-restored fpsr in fpu_from_packed (its only effect was
  already covered by pre_fpsr), SUB_DN_EA's unused src.
- cpu_68000.c and ppc_fpu.c declared a log category and never logged.
- callbacks keep their signatures: (void) on unused parameters.
- scheduler_new_cpu_event_ex: the name lookup that exists only for the
  asserts is compiled out with them (#ifndef GS_FAST).
- cpu_internal.h: two ?: arms made explicitly the same type.

+++ src/core/cpu/cpu_internal.h
@@ -477,7 +477,7 @@ static inline void movem_to_register(cpu_t *restrict cpu, uint16_t opcode, int b
     uint8_t d_set = 0, a_set = 0;
     for (i = 0; i < 8; i++)
         if (register_mask & (1 << i)) {
-            uint32_t v = bits == 16 ? (int32_t)(int16_t)memory_read_uint16(ea) : memory_read_uint32(ea);
+            uint32_t v = bits == 16 ? (uint32_t)(int32_t)(int16_t)memory_read_uint16(ea) : memory_read_uint32(ea);
             ea += bits >> 3;
             if (g_bus_error_pending)
                 return;
@@ -486,7 +486,7 @@ static inline void movem_to_register(cpu_t *restrict cpu, uint16_t opcode, int b
         }
     for (i = 0; i < 8; i++)
         if (register_mask & (0x100 << i)) {
-            uint32_t v = bits == 16 ? (int32_t)(int16_t)memory_read_uint16(ea) : memory_read_uint32(ea);
+            uint32_t v = bits == 16 ? (uint32_t)(int32_t)(int16_t)memory_read_uint16(ea) : memory_read_uint32(ea);
             ea += bits >> 3;
             if (g_bus_error_pending)
                 return;
@@ -1255,7 +1255,7 @@ static inline void f_trap(cpu_t *restrict cpu) {
         return;
     }
     uint32_t fetch_page = cpu->instruction_pc >> PAGE_SHIFT;
-    if (__builtin_expect(g_active_read && (int)fetch_page < g_page_count && g_active_read[fetch_page] == 0, 0)) {
+    if (__builtin_expect(g_active_read && fetch_page < g_page_count && g_active_read[fetch_page] == 0, 0)) {
         // Instruction page has no SoA entry — fetch returned $FF from unmapped
         // physical memory.  Treat as bus error (matching real hardware behavior).
         cpu->pc = cpu->instruction_pc;
